procps: Add append for configuring sysctl
Add bbappend file to procps. This append file allows hardening sysctl
configurations. The sysctl configuration is appended to the end of the
default configuration as this method works for both systemd and sysvinit
systems. Sysvinit does not have sysctl.d directory into which drop-in
configurations could be installed.

Note that user.max_user_namespaces does not seem to work on systemd, as
the value is still non-zero after boot. However, if the namespace
amount is set to 0 after the initialization has finished, it stays zero.
Perhaps systemd initialization touches this value.

This sysctl configuration is added to meta-sulka-kernel instead of
meta-sulka-distro because the sysctl configuration is somewhat tied to
kernel configuration, and is done as a part of the kernel hardening
effort.