linux-yocto: Add development feature
Add a development configuration feature that gets enabled if sulka
development mode is enabled. This allows easing some hardening settings
for development & debugging purposes.

The inital version of the configuration feature allows disabling
SELinux. This helps debugging the system, as SELinux can sometimes be
quite intrusive.

Note that this feature should be preferrably the last in the kernel
features list to prevent it from getting overridden.