linux: Move kernel metadata to a bbclass
Instead of having the kernel metadata under the Linux recipe, move it to
a bbclass. This makes it easier to integrate the kernel hardening
metadata into a different kernel. The users should be able to inherit
the new sulka-kernel-hardening class, instead of having to manually
modify the SRC_URI, KERNEL_FEATURES and so on.